packaging: spec and manifest updates for pmdadm dmstats
RHBZ #1569854

the DM PMDA build requires libdevmapper.h and library in order
to support dmstats metrics. For a spec build, this needs to be
pulled in via BuildRequires otherwise the PMDA will be built
without dmstats support. It looks like device-mapper-devel is
consistently named on all RPM based distros so we don't need any
conditional goo.

For Makepkgs builds, we rely on qa/admin/check-vm to tell us
which package needs to be installed for the build on different
platforms, so also update the package manifest for this.

modified:   build/rpm/fedora.spec
modified:   qa/admin/package.manifest