pmdadm and QA: log a message if HAVE_DEVMAPPER is not defined
If HAVE_DEVMAPPER is not defined, then the PMDA was not built with
libdevmapper.h from device-mapper-devel and hence dmstats metrics
will not be available. Log a warning if this is so.

Update qa/657 with a few additional _notrun scenarios and also
grep the log for the above warning message after installing the
PMDA.

modified:   qa/657
modified:   qa/657.out
modified:   src/pmdas/dm/pmda.c