src/libpcp/src/instance.c: error path memory leak in pmGetInDom()
Very unlikely combination of error conditions, but ...

I think this is the root cause of the Coverity issues (e.g. CID 325933)
over in poor odl pmie's pragmatics.c ... here the caller gets tagged
because pmGetInDom() was not quite doing the right thing.  We'll see
at the next Coverity run after this gets merged.