src/pmlogrewrite: fix error handling when duplicate output files found
We were just removing the output file!  If the output file was a
perfectly good and existing archive, then user command line error
becomes rm -f ... which is not good.

Simple case is when at least one of .index, .meta or .0 files exist.

More complex case is when these ones do not exist so the initial
output archive create works, but some existing volume N (N > 0)
is found to exist later.