pmcd_wait: default to using unix domain sockets if supported
There is a bit of a design flaw in using the "local:" hostspec
for pmcd_wait in that it'll immediately switch away from using
the unix domain socket (to localhost) to connect to pmcd, when
pmcd is initially unavailable.  However that's quite often the
expected situation when pmcd_wait is being used, as we want to
wait for pmcd to come up.

Falling back to the (inet/ipv6) localhost is problematic when
authentication is mandatory (e.g. if pmcd is configured that
way, or if it is configured to disallow certain users/groups
access) because we lose the automatic authentication provided
by unix domain sockets.