rc scripts: improve systemd interaction for local pmie
Add pmie.pid tracking for the primary pmie like the primary
logger, and use this in the systemd service.  Update the
pmcd PMDA to report primary pmie in the same way as it does
the primary logger already, and update pcp-summary(1) to
use this as well.

Test qa/1297 exercises pmie systemd service with systemctl.

Resolves Fedora BZ #1650999.