pcp-dstat: implement the --full option to expand instances
Part of the implementation of -f/--full in the new dstat was
overlooked.  This commit fills in the blanks and extends test
1187 to exercise it.

The code also tweaks -o handling in the case of /dev/stdout
being used as output target.  We cannot seek on anything that
is not a regular file, so check for that before attempting to
append to an existing file.

Resolves Fedora BZ #1651536