libpcp: be more defensive in pmMergeLabelSets
PCP GH#557

Adds a defensive check and pmDebug.labels diagnostic if any element in
the argment labelset is NULL, or has a NULL label pointer.

This occurs with the repro for GH#557 and thus avoids the segfault,
though the root cause in that repo is probably incorrect ctype
casting or invalid argument handling in the pmMergeLabelSets()
python binding. The actual fix is in the python binding and is
still work in progress (only the segfault is avoided so far).
Will need a QA test too, once we have the fix.

modified:   src/libpcp/src/labels.c