qa: make test 1545 more reliable
Marko noticed unreliable results with test 1545 sometimes
failing due to too few lines in the pmrep output.  Turned
out to be an incorrect assumption in the test, when pmrep
is a little slower to start than usual - this results in
the names-change notification arriving on the first fetch
(instead of later on as originally expected).