python pmapi binding: rework labels lookup, merge and free functions
PCP GH#557 - Python PMAPI Labels Segfault

This fixes the segfault reported in GH#557 and also reworks the python
bindings for pmLookupLabels, pmMergeLabelSets, pmFreeLabelSets and the
pmGet*Labels functions. The native C functions generally take an int ID
parameter (e.g pmid or indom, domain, etc) and the address of a pointer
to a pmLabelSet, and return the number of elements in the dynamically
allocated array of pmLabelSets. For the python binding, these functions now
consistently return a python list of pmLabelSets, which is easily iterable.

As a convenience, the first argument to pmMergeLabelSets() can be either
a single pmLabelSet, or a list. In the case of a list being passed in,
each entry may also be a list - this is handy when passing in a list of
the separate results from assorted pmGet*Labels calls, which are now
consistently themselves lists.

The test case in GH#557 will appear in a QA test shortly. The relevant man
pages are also updated with a new PYTHON SYNOPSIS section along with some
examples (next commit).