libpcp, qa: make pcp.conf search algorithm match expectations
Some slight inconsistencies with the documented plan happened
recently; make the code clearer by abstracting pcp.conf search
logic into its little helper (keeping the comment with it) and
fixing a homebrew path inconsistency while there.

There's also a deadlock in the pmNoMem handling, as documented
later in the __pmconfig code, make a note and change handling
of that case to match the way its handled later on (fprintf).

This fixes the issues in qa tests 488 and 963.