pmdapostgresql: ship config that works with a default postgres install
Recent changes added 'osuser' to the default pmdapostgresql
config but left it commented out. This adds it back as the
default user for the PMDA to run as, and thus now works with
a default postgresql server installation where the postgres
user can access the database over unix domain connection.

modified:   src/pmdas/postgresql/pmdapostgresql.conf