qa: towards correct pmcd stop/start behaviour with unix sockets
There is an oddball case stopping/starting pmcd since we made
pmcd_wait default to waiting on the unix domain socket when it
is supported.  This is where -s/--socket is used with pmcd and
its not in the usual location ... we then do not know where to
find the socket path to wait on.

This is the first changes toward qa/956 working, more to come.