pmlogger_check: make pmlogconf check optional
Since the dawn of time, QA has been plagued by failures with the
signature:
Arrgghhh ... pmlogger (primary) failed to start after 20 seconds

Finally tracked this down to the slowness of pmlogconf on some VMs
and the fact that we're running pmlogconf from pmlogger_check for
no useful purpose most of the time.  On one machine (vm33) I clocked
pmlogconf taking 25 seconds to redo the configuration, which is
longer than the 20 seconds _wait_for_pmlogger() is willing to hang
around.

From the (updated) pmlogger_check man page ...

    The default behaviour, when pmlogger(1) configuration comes from
    pmlogconf(1), is to regenerate the configuration file and check for
    changes whenever pmlogger(1) is started from pmlogger_check. If the
    PMDA configuration is stable, this is not necessary, and setting
    $PMLOGGER_CHECK_SKIP_LOGCONF to yes disables the regeneration
    and checking.

PMLOGGER_CHECK_SKIP_LOGCONF can be set in the environment, or
for a bigger hammer approach, in the $PCP_SYSCONFIG_DIR/pmlogger
defaults file.  But be careful that setting it in the latter means
config.default will never be regenerated even if a PMDA is installed
or removed, or a new PCP version is installed.