qa: add PMLOGGER_CHECK_SKIP_LOGCONF=yes for assorted tests
These are the ones that in the recent past have cause QA failures
when not restarting pmlogger within the 20 second timeout.

We're taking an informed punt (based on non-determinism) that the
root cause is the time taken to run the pmlogconf reconfiguration,
and if so setting PMLOGGER_CHECK_SKIP_LOGCONF=yes should help.