docs: update pmLookupLabels(3)
Update pmLookupLabels(3) to document that the returned array of label sets
no longer includes the instances level labels. For the python binding,
the returned value is a list of pmLabelSet class instances, excluding the
instances level label sets.

Also add a new section "PYTHON EXAMPLE", with some example python code
for retrieving and printing the merged label sets for a particular metric,
and (separately) for the instances level labels if that metric has a
non-singular indom.

modified:   man/man3/pmlookuplabels.3