src/pmlogger/pmlogger_rewrite.sh: sum(1) is not always there!
For *BSD systems it is more likely to be the (Posix compliant)
cksum(1) that is present ... check for both and use sum(1) (so minimal
functional change from the historical behaviour) if available, else
cksum(1) if available, else wc(1) (which does not really work as
a checksum).

This turns out to be benign here as we're only checking before
optionally recompressing, sum(1) and cksum(1) do the same job.