src/libpcp_fault/src/GNUmakefile: change rules for derive_parser.y
In libpcp_fault, mimic the libpcp rules so that derive_parser.y.in
is a symlink back to libpcp and then make derive_parser.y locally
for the build.  Before this, we could not edit derive_parser.y.in in
the libpcp_fault/src directory and have the changes reflected back
in the libpcp/src directory.
[this is only an issue for Ken!]