pmseries: incorporate feedback and fix bugs from Marko
Several changes in pmseries behaviour here.  In the presence
of no command line options, instead of assuming the arguments
form a query, we now use a heuristic to determine if we might
be looking at timeseries identifiers.  This seems to be quite
a common assumption, so embrace it.

Fixed a couple of bugs Marko found around error handling - if
a non-existent archive is presented, fail early before trying
to use it.  Fix an issue with labels handling which caused a
connection shutdown before processing all request results and
also harden the callback handlers.