pmproxy pmda: extend MMV to support pmproxy metric dog-fooding
This is a series of changes to the MMV agent allowing it to run
as a pmdapmproxy as well, sourcing pmproxy.* metrics from a new
(alongside existing MMV) filesystem location.  This means we do
not run the risk of the user permission modifications for a MMV
system mapping directory affecting pmproxy.

The slightly difficult part of this work involved adding a data
pointer for shared PMDAs to use (cannot use global variables in
this PMDA as we used to, the two clobber each other).  This is
accessible via some helper routines in libpcp_pmda now, as has
been documented in pmdaInit(3).

IOW, the same shared library (DSO) supports both "agents", and
pmproxy is a default PMDA (alongside pmcd, MMV, kernel), so the
metrics are always available.  Incidentally, this means we will
now get more coverage of MMV as we're dog-fooding it instead of
relying solely on QA and end-users using it.

Simple test case added; instrumentation of pmproxy, its modules
and libpcp_web will follow.