pmlogconf: add per-CPU interrupts metrics to the zeroconf set
In some investigation cases, we need /proc/interrupts information
to analyze issues from the following viewpoint:
  - Whether specified IRQs are distributed across processors in
    an appropriate NUMA node as we expected.
  - What IRQ causes a CPU to be high usage.

This change makes zeroconf collect kernel.percpu.interrupts by default.

We set "delta 10 minutes" in the patch, because we think it is enough
interval as for the metric, and because the size of archive tends to be
large if /proc/interrupts is collected completely and frequently due to
its n_IRQs rows and n_CPUs columns matrix information.