Fix Python install flags for DIST_ROOT=
Fix an invalid shell variable reference in builddefs.in which
accidentally works when the DIST_ROOT variable is not empty, but causes
installation of Python modules to fail when DIST_ROOT is empty.  The
Homebrew packaging system for MacOS uses an empty DIST_ROOT to build and
install directly into a user-writable directory under /usr/local/Cellar.