pmcd: allow use of an optional pmcd.conf.access file
To aid people wanting to manage access control sections
in pmcd.conf files under automated control, allow for a
separate file with access control specifications.  This
file is optional; but if no pmcd.conf [access] section
then it is solely used for access specification.

This is a backward-compatible change, and we expect to
continue to ship pmcd.conf with an [access] section for
eternity.  For anyone wanting to use only .access files

  $ sed -i '/^\[access\]/,$d' /etc/pcp/pmcd/pmcd.conf

Test 546 is updated to check syntax, and new test 1511
is added to verify it behaves as advertised.

This resolves github issue #611.