Rename pmconfig.py helper function
get_sorted_results() is somewhat confusing name as man pages and
code elsewhere always refers to ranking, not sorting.

Rename the helper function to clarify its workings a bit (and to
match its comments) but leave a backward-compatible wrapper for
possible out-of-tree clients for the time being, could be removed
e.g. at PCP 5.0 or so. Update all in-tree callers.