pmproxy: fix shutdown issues in timeseries mode
Marko reported long delays in shutting down pmproxy when
running in timeseries mode.  Turns out to be a result of
intermixing libuv use with regular signal handling.  So,
switch to using libuv handlers which integrate with the
async event loop, and signals are delivered correctly.

Several other issues come up in testing this relating to
having outstanding requests in flight, then cancelled -
these caused assertion failures in libuv.

Add test qa/1433 to exercise this all and verify we have
a clean pmproxy log; a common.filter routine for pmproxy
log filtering is added to help.