systemd: rearrange timers to not use sysconfig settings
We found in practice the systemd mechanisms for setting
"environment" variables to be too brittle for our needs
so revert to simpler unit files without sysconfig files.

The problems related to the way systemd replaces those
variables when no values, or multiple shell parameters,
were needed - passing these in as a single shell option
(possibly empty) was not what anyone expected.