systemd: ensure timers and services get started together
Found a case where some (pmie_check, etc) of the auxillary
services were not getting automatically started alongside
pmie/pmlogger service, via systemd.  From further reading
the BindsTo is not enough alone, and must be used with the
After keyword to ensure the behaviour we need.  Make It So.

While there, fix the man page references for some of the
auxillary services.

Related to https://github.com/performancecopilot/pcp/pull/599