src/libpcp_pmda/src/open.c: don't timeout waiting for credentials from pmcd
If pmcd is busy (servicing other clients and/or PMDAs), just be patient
rather than timing out if the credentials PDU does not turn up in
a few seconds.  The PMDA has nothing to do until the credentials
exchange happens, so no badness from that quarter if we wait a bit
longer, but before this change we were seeing some QA failures on
slow VMs with PMDA timeouts at this PDU exchange.