scripts/groups-to-test: map git commits onto QA tests and test groups
Lukas mentioned at the Melbourne PCP Conference that it would be useful
(in the context of moving to a QA-before-merge workflow) to be able to
run just the QA tests covering the code changes in a set of git commits.

I said this "should be possible" ... many hours and 800+ line of
shell script later, scripts/groups-to-test does the job.

It uses git-diff (git-show is not reliable for our git tree, possibly
due to some renaming) to generate a list of the names of all the files
that are changed (modified, added or deleted) between two commits (or
one commit and the current state of the tree), and then maps these
file names onto a set of QA groups and/or QA tests.  The output is
an argument string that could be passed to qa/check.