Allow --with-qt=DIR to specify QT installation dir
The configure script cannot be given a value for the QT installation
directory except by setting the environment variable QTDIR.  There is a
--with-qt option but you cannot pass it a directory, the only valid
values are "yes" "no" and "check".

Regrettably Homebrew makes it super hard to set an environment variable
during build, and really wants you to do everything with commandline
options to configure.

This change extends the syntax of the --with-qt option by allowing its
value to be the absolute path of the QT installation directory.  This is
consistent with many open source users of autoconf.