pmcd: add support for PMDA suspend and resume (fencing)
This addresses a (Red Hat HPC customer) need to suspend and resume
agents, to allow fine grained parallel applications to run without
interruption.  As states in the original request: "overall compute
efficiency goes down drastically when these kind of jobs which need
to interact with each other get interrupted".

This extends a little on the prototype that I described earlier -
https://groups.io/g/pcp/message/19958 - in that it allows per-PMDA
suspend/resume.  This was done after feedback on the prototype was
received, and it turned out to be slightly cleaner to implement in
the end too.  In this version, there is no pmfence(1) command - we
are able to make use of the existing pmstore(1) command.

A new section is added to pmcd(1) titled "AGENT FENCING" describing
how to operate this functionality.  Regression test qa/1340 is also
added to exercise the code.

This resolves Red Hat BZ #1641161