pcp2zabbix: send first metrics without waiting
Rather oddly, pcp2zabbix was not reporting (sending) the first set
of metrics to the Zabbix server prior pausing (as per -t) like all
the other pmrep/pcp2xxx tools. There seems to be now logical reason
for this so we assume an oversight and change pcp2zabbix to be like
the rest of the tools. Even QA seems to content with this change.

Additionally, better handle Ctrl-C from the user while attempting
to connect to the server.