bcc PMDA: remove PM_ERR_AGAIN uses
This is the first of a series of PMDA audits and (largely cosmetic)
changes.

Specifically we're looking at the fetch callback method in the PMDA
which should return PMDA_FETCH_NOVALUES (rather than PM_ERR_AGAIN)
for cases where metric values are not currently available and there's
no good reason to think they might become available any time soon
(for a particular platform and software environment).

For some PMDAs (not the bcc one) this will also mean updating from
PMDA_INTERFACE_2 to one of the more recent PMDA interfaces to get
the 2-state or (better) 3-state fetch callback semantics.