qa bcc tests: stamp QA version of bcc.conf
Protocol (not written down anywhere) is that QA tests that change
configuration files from packages should include the annotation

    Installed by PCP QA test $seq on `date`

as a comment to help diagnose errant QA tests that muck with
configuration files and don't clean up properly.

There is no suggestion that the bcc QA tests are not cleaning up
properly, this is just a precaution.