qa/check et al: rework callback workflow
We used to (optionally) run check.callback after each QA test to look
for pmcd and pmlogger badness.

This has been extended to capture the state of the pmcd and pmlogger
landscape before each QA test is run, and then if the post-test
callback detects a problem report the pre-test state along with the
post-test failure.

It is particularly useful to know which pmcd and pmlogger pids are
running before and compare this with after to help diagnose what
went wrong.