qa: massive change to remove init script chatter
The chatter from our init scripts like ...

    Starting pm*
    Stopping pm*
    Waiting for pm*

do not help test pass/fail as the tests always do something after
the stop/start/restart to check the associated daemons are or are
no running.

But more critically, when systemctl is in the mix, these messages
are salted away with journalctl (or equivalent) and have proven to
be incredibly difficult to extract reliably (buffering, overrun with
other messages, multiple instances of the messages in the recent past,
etc.).  If our QA tests are not dependent on this chatter appearing,
then we can re-consider using systemctl in QA.