src/libpcp_pmda/src/callback.c: guard use of e_labelCallBack callback
If using INTERFACE_7 and pmdaSetLabelCallBack() is not called, then
e_labelCallBack is NULL and we SEGV when trying to make the callback.
Add a guard to prevent this.

Observed with the mounts PMDA (recently converted to INTERFACE_7) and
qa/799 on bozo-vm where the mounts PMDA is permanently Installed
(for reasons known only to god, and she's not sharing why).