pmproxy: rework the way JSON output termination is performed
The existing mechanism using flags became clunky as soon as
non-trivial object/array nesting was needed.  Switch to use
of a stack of object/array termination characters.

Also, flesh out the series/labels output to knock off a code
TODO item - updating test 1601 to reflect the new expected
result.

Finally, update a couple of pmproxy/pmseries tests to take a
bit more care with ensuring configuration files are not in a
position to influence test results.