SELinux changes: allow pmdasimple to chown named pipe when Enforcing
This was causing qa/110 to fail on platforms where SELinux was in
Enforcing mode, because this use of the simple PMDA uses a named
pipe in /var/tmp (or /tmp) and then tries to chown it, which fails
and leads to badness in pmda-pmcd land.

In hindsight it would probably have been better to change qa/110 to
create the named pipe someplace SELinux was less sensitive about, but
I used the opportunity to exercise and refine Lukas' notes on how
to make these sort of SELinux changes.