libpcp_web: refactoring and support for Redis cluster protocol
Add support for the MOVED and ASK responses to Redis commands
when operating in a Redis cluster.  The commands allow dynamic
changes to the cluster host set and are mandatory.

In order to support this, we need to keep track of the commands
we submit during a request, so that redirection to an indicated
slot host can be performed when necessary (this is via special-
casing of error responses).  Command redirection is achieved by
attaching the command sds to every request at a relatively low
level, and passing it back with the callback.

Existing pmseries QA passes with both single Redis instance and
with clusters now.