src/pmlogger/pmlogger_daily.sh: try harder in the presence of errors
When trying to process archives for a specific day, there are a number
of possible error conditions that may occur.  Historically, we viewed
these errors as something that required human intervention, and as a
consequence tended to bail as soon as a problem is encountered.

Fast forward, and most pmlogger farms are probably unattended these
days, and the "bail early" strategy means that once a problem occurs
we may keep seeing the same problem and fail to process subsequent
archives ... in the extreme cases this may mean no merging and
no compressing.

This change means that errors are reported as before, but we continue
to process any archives for any other days and hosts.