qa: add selinux specials
For some QA tests we're getting AVC errors with selinux due to the
QA test applications or the way we're using PCP applications ...

This commit adds a new selinux module (pcpqa) that is only installed
if:
(a) selinux is installed, and
(b) selinux is Enforcing, and
(c) the pcpqa module is not already installed

There is also some associated fiddling with PCP-specific selinux
"booleans" when the pcpqa module is installed.

Note AVC errors during QA that are _really_ generic PCP ones, need
to be fixed in ../src/selinux.  The pcpqa module here is just for
extra capabilities needed by the unusual demands of QA.