qa/admin/pcp-daily: conditionally run freespace
In the QA Farm, about 10% of my runs (and much higher after a PCP
version change) are aborted because I run out of disk space on the VMs
(because I was too mean when they were provisioned).

I've been using a freespace script manually for some time and this
commit just automates the excution of that script if it exists.