qa/451 & qa/998: clean up any core file
These tests are using libpcp_fault and drive the code down the error
paths that end in abort() ... if the system is configured to allow
core dumps, we get a core dump and check thinks the test has failed.

Remove the core file if any, as this is expected and not an indication
of a test failure.