qa: many: purge explicit $seq.notrun uses and replace by _notrun()
When using check.callback.sample with the new AVC checks, if
a test is "notrun" and uses the old protocol of explicitly
creating a $seq.notrun file, rather than calling _notrun()
we end up with $seq.pre-avc turds in the QA directory.

Converting all these tests to use _notrun() avoids the problem
as there is a cleanup rm there for $seq.pre-avc if it exists.