logutil: change pmlogger_daily_report to be an opt-in service
Change the pmlogger_daily_report service (part of pcp-zeroconf)
to be "opt-in", rather than enabled by default, with the current
setting retained on upgrades. When run, the script invokes pmrep(1)
multiple times, which can be expensive because the input archive(s)
are transparently decompressed each time but there is no caching of
recently decompressed archives.