libpcp_web: Solaris changes
The conditional inclusion of sslio.c, libuv.c, discover.c and (most
importantly) webgroup.c does not happen on Solaris (vm13), so the
exports file is wrong and the dso build has warnings/errors.

Refactor to make the exports file on the fly (from exports.in) based
on which source files are going to be included.