pmlogextract: add -x option for more resilient merging
For QA machines in particular, the varied use of the mmv PMDA may mean
the daily archives contain metrics with mismatched metadata, then
pmlogextract chokes in the nightly pmlogger_daily script.  Adding -x
to pmlogextract will silently "fix" this.

It may also be useful in unattended pmlogger environments.

From the man page ...
   -x     It  is  expected  that the metadata (name, PMID, type, semantics
	  and units) for each metric will be consistent across all of  the
	  input  PCP archive log(s) in which that metric appears.  In rare
	  cases, e.g. in development, in QA and when a PMDA  is  upgraded,
	  this  may not be the case and pmlogextract will report the issue
	  and abort without creating the output archive log.  This is done
	  so the problem can be fixed with pmlogrewrite(1) before retrying
	  the merge.  In unattended or QA environments it may  be  prefer‐
	  able to force the merge and omit the metrics with the mismatched
	  metadata.  The -x option does this.

Add qa/813 to test this.

Also a bunch of code cleanups in pmlogextract.

And correction of a corner case bug that has been there forever ... if
you use -c configfile and have a non-leaf PMNS name in configfile and
that non-leaf expands to different metrics across the input achives,
then only the matching metrics from the first archive containing
the non-leaf PMNS name would be included in the output ... the test
archives for qa/813 triggered just this scenario!
Fixed.