pmdalinux: fix hugepage metric value calculations
The values for some of the Linux kernel hugepage metrics were
being incorrectly multiplied by 1024.  This was due to a code
assumption in proc_meminfo.c::refresh_proc_meminfo around line
115, which was assuming all meminfo values needed conversion
from kbytes to bytes - not so for these metrics.

The fix involves removing this assumption (for all values from
/proc/meminfo) and individually applying unit conversion where
needed only.  Updated calculations are now reflected in qa/821.

Resolves Red Hat BZ #1730107.