pmchart: fix chart auto-scaling under fetch error conditions
pmchart requires the data series bounds to be accurately calculated
to rescale the plot axis dynamically. When a metric fetch fails, the
sample value in the data series is set to NaN, indicating that the
sample is not valid.

When we calculate the bounding rectangle of the data series, we
should skip over these invalid samples as they are by definition,
aren't finite and are not plotted. Hence they should not contribute
to the data series bounding box. Doing so causes issues with (at
minimum) line chart rescaling as NaN is considered to be both
greater than the maximum finite value and less than the minimum
finite value, resulting in the height of the Y-axis being NaN. And
because the Y-axis height is not finite, the line charts do not
rescale.

Hence while there is a NaN sample in the data series, line charts
will not automatically rescale. Because the first sample in a live
data series is almost always a metric fetch error, line charts won't
rescale correctly at startup. And when multiple hosts are being
monitored live and are being regularly rebooted, there is always
missing samples in every data series, and hence line charts almost
never rescale.

Fix this by ignoring bounding boxes for data samples that result in
NaN bounds when recalculating the bounding rectangle of a data
series. This results in the bounding box range spanning only the set
of valid samples.

Signed-off-by: Dave Chinner <dchinner@redhat.com>