build: rework libpcp_web exports file exposed-symbols
Rethink the approach to handling libpcp_web symbols that are
not available due to missing platform features like libuv.
The previous method failed when a single optional symbol was
present in a symbol VERSION group (that group is then empty
on some platforms, and syntax errors result).

Instead, present a single ABI always and use stub routines
(noops or returning errors) for unavailable features.