pmdaproc: fix memory leak in pidlist refresh
RHBZ #1721107

The pidlist refresh function was leaking instance name buffers for
processes that had exited between refreshes. Fix this by storing the
instance name (truncated psargs) in a new field of the hash table
entry, and harvest them correctly if they've exited after each refresh.
The indom table instance names now share a pointer to the new field
in the hash table entry so they dont need to be separately free'd.
This similifies the code and memory management.

Also put a 4K cap on the maximum length of a psargs string for the
unlikely scenario of a psargs string that is not NULL terminated,
i.e. use strndup() rather than strdup(). This may be possible if
a rogue process modified it's argv[0] at run-time.

For testing, the recipe in RHBZ#1721107 was repro'd using valgind
and a workload with lots of process churn (i.e. a couple of large
builds running). Before the fix, valgrind reported numerous similar
leaks as reported in the bug. With the fix applied, those leaks
are gone. The refresh is more efficient now too because we only
have one copy of the psargs strings (in the hash table entry).

Also thoroughly exercised all tests in the QA pmda.proc and
pmda.hotproc groups. A new test might be possible, but the nature
of the required workload would make it difficult (and expensive)
to run.