libpcp_web: fix discovery segfaults when deleting monitored paths
Deleting an entire monitored directory (as is done at the end of qa/1600)
was traversing the hash table and deleting paths in that directory,
which was racing with other callbacks handing the deletion of individual
archive paths within the deleted directory .. causing double-frees/segfaults.

Deleting a monitored archive meta file was also segfaulting when the
log rotation scripts restarted pmlogger (the metadata file gets
compressed, so the original is deleted).

With these changes, pmproxy/libpcp_web archive discovery is more
stable and qa/1600 is now passing (it was failing with the directory
removal tests at the end of the qa test). There may however be memory
leaks that will require further investigation.

There are still some prevailing pmproxy segfaults involving use-after-free
of batons - these seem to be unrelated to the changes in this commit.