log management - fix broken pipe errors from _is_archive() in utilproc.sh
mkaf, pmafm and other utilities used by pmlogger_daily and other
log management scripts often partially fail with mysterious broken
pipe messages in the logs :

Warnings/errors from mkaf ...
cat: write error: Broken pipe

This seems to be stderr from the _is_archive() shell function in
utilproc.sh, which runs various decompression utils (or cat) on it's
argument, which is then piped into dd to read the first 7 bytes (only)
and then into awk to check for PCP file signatures. This results in
the start of that pipeline exiting with a broken pipe, which sometimes
(it's racy) propagates stderr messages through to the original
_is_archive caller, causing it to think mkaf (or whatever is calling
_is_archive) has failed .. sometimes.

This patch directs stderr to /dev/null after the first stage in the
_is_archive pipeline, and now pmlogger_daily and friends are reliably
happier.