logutil: fix pmlogger and pmlogger_daily systemd services
Fedora BZ #1737091 - pmlogger slows down boot significantly
Fedora BZ #1721223 - pmlogger.service: Failed with result 'timeout'

The systemd configs for pmlogger_daily (both service and timer) should
NOT be required by the pmlogger service - otherwise, when the pmlogger
service attempts to start up, the pmlogger_daily service is started
too. The pmlogger_daily service fails if the pmlogger service is not
running, which then causes the pmlogger service to be terminated (this
is the systemd behaviour with RequiredBy=pmlogger.service).

So we end up with a lot of confusing log messages, timeouts and super
slow boots, and ... usually no pmlogger services running at all!
The previous commit concerning broken pipe messages is related too.

With these fixes, the pmlogger service now takes 6 to 10 seconds to
start (compared to ~ 30 seconds delay during boot and then failing to
start at all). This is still too long, so there will need to be
followup work to improve that for BZ#1737091.