build: add pmdastatsd and dependencies into the build
Initial work to add new toolchain (ragel) and dependent
libraries (hdr_histogram and chan) detection and use to
the build, so pmdastatsd can be built as part of PCP.

Small updates to docs, file permissions, and so on are
also made here.

Much work remains for packaging, QA, etc - but this is
enough to get the code to build.