libpcp: fix cockpit-pcp segv with local context
The problem has variously been reported as:
    https://github.com/cockpit-project/cockpit/issues/6108 and
    https://bugzilla.redhat.com/show_bug.cgi?id=1235962
and most recently associated with
    https://github.com/performancecopilot/pcp/pull/693

The root cause was a corner case with multiple contexts using
PM_CONTEXT_LOCAL and instance profiles.  We had an optimization that
stopped pushing the profile to the DSO PMDA if it had already been
pushed.

The logic around this optimization was wrong, and when combined
with calls to pmDestroyContext() we could end up with a stale
and corrupted instance profile being used in the DSO PMDA.

This change moves the state from the client-side per-context
structure to the PMDA-side __pmDSO structure.